+++ src/com/company/Node.java
@@ -2,12 +2,12 @@ package com.company;
 
 class Node {
 
-    int value;
+    long value;
     Node left;
     Node right;
     boolean onRightSide;
 
-    Node(int value, boolean onRightSide) {
+    Node(long value, boolean onRightSide) {
         this.value = value;
         right = null;
         left = null;
